添加水波纹效果，原帖的点击事件bug修复
