+++ app/src/main/res/layout/home_complete_item.xml
@@ -4,6 +4,9 @@
     android:orientation="vertical"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
+    android:foreground = "?attr/selectableItemBackground"
+    android:clickable="true"
+    android:focusable="true"
     android:layout_marginBottom="5dp"
     android:layout_marginTop="5dp"
 
@@ -13,7 +16,6 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:orientation="vertical">
-
         <include layout="@layout/home_item_head"></include>
