+++ app/src/main/res/layout/home_image_only_item.xml
@@ -4,6 +4,9 @@
     android:orientation="vertical"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
+    android:foreground = "?attr/selectableItemBackground"
+    android:clickable="true"
+    android:focusable="true"
     android:layout_marginBottom="5dp"
     android:layout_marginTop="5dp"
     app:elevation="5dp">
